+++ paramiko_functions.py
@@ -3,8 +3,8 @@ import paramiko
 import sshtunnel
 
 
-def check_host_availability(username, host):
-    pkey = paramiko.RSAKey.from_private_key_file(os.path.expanduser('~/.ssh/id_rsa'))
+def check_host_availability(username, host, ssh_key_path):
+    pkey = paramiko.RSAKey.from_private_key_file(ssh_key_path)
 
     client = paramiko.SSHClient()
 
@@ -15,16 +15,16 @@ def check_host_availability(username, host):
     client.close()
 
 
-def check_target_host_availability(username, jump, host, local_port):
+def check_target_host_availability(username, jump, host, local_port, ssh_key_path):
     with sshtunnel.open_tunnel(
             (jump, 22),
             ssh_username=username,
-            ssh_pkey=os.path.expanduser('~/.ssh/id_rsa'),
+            ssh_pkey=os.path.expanduser(ssh_key_path),
             ssh_private_key_password='',
             remote_bind_address=(host, 22),
             local_bind_address=('0.0.0.0', local_port)
     ) as _:
-        pkey = paramiko.RSAKey.from_private_key_file(os.path.expanduser('~/.ssh/id_rsa'))
+        pkey = paramiko.RSAKey.from_private_key_file(ssh_key_path)
         client = paramiko.SSHClient()
         # Allow hosts that not in known_hosts
         client.set_missing_host_key_policy(paramiko.AutoAddPolicy())
@@ -32,23 +32,16 @@ def check_target_host_availability(username, jump, host, local_port):
         client.close()
 
 
-def forwarding_through_jump_server(username, jump, host, local_port, host_port):
+def forwarding_through_jump_server(username, jump, host, local_port, host_port, ssh_key_path):
     tunnel = sshtunnel.open_tunnel(
         (jump, 22),
         ssh_username=username,
-        ssh_pkey=os.path.expanduser('~/.ssh/id_rsa'),
+        ssh_pkey=ssh_key_path,
         ssh_private_key_password='',
         remote_bind_address=(host, host_port),
         local_bind_address=('0.0.0.0', local_port)
     )
     tunnel.start()
-
-    # pkey = paramiko.RSAKey.from_private_key_file(os.path.expanduser('~/.ssh/id_rsa'))
-    # client = paramiko.SSHClient()
-    # # Allow hosts that not in known_hosts
-    # client.set_missing_host_key_policy(paramiko.AutoAddPolicy())
-    # client.connect('localhost', pkey=pkey, username=username, port=local_port)
-    # client.close()
 
     return tunnel